Move Aspire ServiceDefaultExtensions to SharedInfrastructureConfiguration to make it available to workers too
