+++ application/account-management/Workers/Program.cs
@@ -6,7 +6,7 @@ using PlatformPlatform.SharedKernel.Database;
 // Worker service is using WebApplication.CreateBuilder instead of Host.CreateDefaultBuilder to allow scaling to zero
 var builder = WebApplication.CreateBuilder(args);
 
-// Configure storage infrastructure like Database, BlobStorage, Entity Framework DB Context, etc.
+// Configure storage infrastructure like Database, BlobStorage, Logging, Telemetry, Entity Framework DB Context, etc.
 builder
     .AddDevelopmentPort(9199)
     .AddAccountManagementInfrastructure();
